+++ test/jdk/java/security/KeyPairGenerator/Failover.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2003, 2016, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2003, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -24,7 +24,7 @@
 /**
  * @test
  * @bug 4894125 7054918 8130181
- * @library ../testlibrary
+ * @library /test/lib
  * @summary test that failover for KeyPairGenerator works
  * @author Andreas Sterbenz
  */
@@ -32,8 +32,8 @@
 import java.util.*;
 
 import java.security.*;
-import java.security.interfaces.*;
 import java.security.spec.*;
+import jdk.test.lib.security.ProvidersSnapshot;
 
 public class Failover {
