+++ test/jdk/java/security/Provider/GetInstance.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2003, 2016, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2003, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -24,7 +24,7 @@
 /*
  * @test
  * @bug 4856968 7054918 8130181
- * @library ../testlibrary
+ * @library /test/lib
  * @summary make sure getInstance() works correctly, including failover
  *   and delayed provider selection for Signatures
  * @author Andreas Sterbenz
@@ -34,6 +34,7 @@ import java.util.*;
 
 import java.security.*;
 import java.security.cert.*;
+import jdk.test.lib.security.ProvidersSnapshot;
 
 public class GetInstance {
