+++ test/jdk/java/security/Security/NoInstalledProviders.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1999, 2011, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1999, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -24,12 +24,13 @@
 /*
  * @test
  * @bug 4273454 7054918 7052537
- * @library ../testlibrary
+ * @library /test/lib
  * @summary Make sure getProviders(filter) doesn't throw NPE
  * @run main/othervm NoInstalledProviders
  */
 
 import java.security.*;
+import jdk.test.lib.security.ProvidersSnapshot;
 
 public class NoInstalledProviders {
