+++ test/jdk/java/security/cert/CertPathValidator/trustAnchor/ValWithAnchorByName.java
@@ -52,8 +52,8 @@ public class ValWithAnchorByName {
 
     // The following certificates and OCSP responses were captured from
     // a test run that used certificates and responses generated by
-    // sun.security.testlibrary.CertificateBuilder and
-    // sun.security.testlibrary.SimpleOCSPServer.
+    // jdk.test.lib.security.CertificateBuilder and
+    // jdk.test.lib.security.SimpleOCSPServer.
 
     // Subject: CN=SSLCertificate, O=SomeCompany
     // Issuer: CN=Intermediate CA Cert, O=SomeCompany
