+++ test/jdk/javax/crypto/EncryptedPrivateKeyInfo/GetKeySpecException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2003, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2003, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -24,19 +24,20 @@
 /**
  * @test
  * @bug 4508341 7055362
- * @library ../../../java/security/testlibrary
+ * @library /test/lib
  * @summary Test the error conditions of
  * EncryptedPrivateKeyInfo.getKeySpec(...) methods.
  * @author Valerie Peng
  * @run main/othervm -DcipherAlg=PBEWithMD5AndDES GetKeySpecException
  * @run main/othervm -DcipherAlg=PBEWithSHA1AndDESede GetKeySpecException
  */
+
 import java.security.*;
-import java.util.Arrays;
 import java.util.Vector;
 import java.security.spec.*;
 import javax.crypto.*;
 import javax.crypto.spec.*;
+import jdk.test.lib.security.ProvidersSnapshot;
 
 public class GetKeySpecException {
     private static String cipherAlg;
