+++ test/jdk/javax/crypto/JceSecurity/SunJCE_BC_LoadOrdering.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2006, 2011, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2006, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -24,7 +24,7 @@
 /*
  * @test
  * @bug 6377058 7055362
- * @library ../../../java/security/testlibrary
+ * @library /test/lib
  * @summary SunJCE depends on sun.security.provider.SignatureImpl
  * behaviour, BC can't load into 1st slot.
  * @author Brad R. Wetmore
@@ -33,7 +33,7 @@
 
 import java.security.*;
 import javax.crypto.*;
-import java.io.*;
+import jdk.test.lib.security.ProvidersSnapshot;
 
 public class SunJCE_BC_LoadOrdering {
