+++ test/jdk/sun/security/tools/keytool/KeyToolTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2005, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2005, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -23,7 +23,7 @@
 
 /*
  * @test
- * @library /java/security/testlibrary
+ * @library /test/lib
  * @bug 6251120 8231950 8242151
  * @summary Testing keytool
  *
@@ -68,9 +68,9 @@ import java.security.NoSuchAlgorithmException;
 import java.util.*;
 import java.security.cert.X509Certificate;
 import jdk.test.lib.util.FileUtils;
+import jdk.test.lib.security.HumanInputStream;
 import jdk.test.lib.security.SecurityUtils;
 import sun.security.util.ObjectIdentifier;
-
 
 public class KeyToolTest {
