+++ test/jdk/sun/security/x509/URICertStore/AIACertTimeout.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2023, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -27,8 +27,7 @@
  * @summary Enhance OCSP, CRL and Certificate Fetch Timeouts
  * @modules java.base/sun.security.x509
  *          java.base/sun.security.util
- * @library /test/lib ../../../../java/security/testlibrary
- * @build CertificateBuilder
+ * @library /test/lib
  * @run main/othervm -Dcom.sun.security.enableAIAcaIssuers=true
  *      -Dcom.sun.security.cert.readtimeout=1 AIACertTimeout 5000 false
  * @run main/othervm -Dcom.sun.security.enableAIAcaIssuers=true
@@ -55,10 +54,10 @@ import java.security.spec.*;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 
-import sun.security.testlibrary.CertificateBuilder;
+import jdk.test.lib.security.CertificateBuilder;
 
-import sun.security.x509.*;
-import sun.security.util.*;
+//import sun.security.x509.*;
+//import sun.security.util.*;
 
 public class AIACertTimeout {
